+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 
 ## Added
-- An `evaluate_stokes` function that will return a new SkyModel at requested frequencies.
+- An `at_frequencies` function that will return a new SkyModel at requested frequencies.
 - A new `component_type` parameter that can be set to "healpix" or "point".
 - Better support for HEALPix maps, with new `nside` and `hpx_inds` parameters.
 - Added the following methods to the `SkyModel` class: `select`, `source_cuts`,